moved tx creation on proposals to a component
