+++ components/Proposal.tsx
@@ -6,23 +6,13 @@ import { Cw3FlexMultisigNamedClient } from 'hooks/guildapp-ts/Cw3FlexMultisigNam
 import { useSigningClient } from 'contexts/client';
 import { Coin } from 'coreum/proto-ts/cosmos/base/v1beta1/coin';
 import { tokensList } from 'util/constants';
-import { MsgIssue } from 'coreum/proto-ts/coreum/asset/ft/v1/tx';
-import { FormControl, FormLabel, FormGroup, FormControlLabel, Checkbox } from "@mui/material"
-import { MsgSend } from 'coreum/proto-ts/coreum/nft/v1beta1/tx';
-import { AssetFT  } from 'coreum/tx'
-import { MsgExecuteContract } from 'hooks/generated-ts/cosmwasm/wasm/v1/tx';
-//import {
-//  MsgInstantiateContract,
-//  MsgExecuteContract,
-//  Msg
-//} from '../hooks/coreum-ts/asset/ft/v1/tx';
+import ProposalTxCreator from './proposalTxCreator';
   
 /* 
 TO FIX:
 - continue FT issuance
-
-- Add tx generator
-- handle response 
+- NFT (issuance & transfers)
+- handle responses 
 
 
 */
@@ -37,101 +27,17 @@ const Proposal: NextPage = () => {
     msgs: [],
   });
 //    action: "", // create || transfer (later could be added "mint, burn, freeze, unfreeze")
-  const [txIssueForm, setTxIssueForm] = useState({
-    subunit: '',
-    symbol: '',
-    precision: 18,
-    initialAmount: "0",
-    features: [1, 0],
-    sendCommissionRate: '0'
-  })
-  const [txTransferForm, setTxTransferForm] = useState({    // transfers only
-    type: "",   // ft || nft
-    token: '',  // token denom
-    amount: 0,
-    destination: '',
-  })
   const [msgs, setMsgs] = useState([])
   const [txModel, setTxModel] = useState<string>('none')
   const [success, setSuccess] = useState<string>('')
-  const [vaultSelected, setVaultSelected] = useState()
+  const [vaultSelected, setVaultSelected] = useState<string>('')
   const [vaultBalance, setVaultBalance] = useState<Coin[]>([])
-  const [vaultFtSelected, setVaultFtSelected] = useState<Coin | null>(null) 
-  const [issueTokenFeatures, setIssueTokenFeatures] = useState({
-    mintable: true,
-    burnable: false,
-// ..more features of smart tokens
-  })
-  const handleFeaturesChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setIssueTokenFeatures({
-      ...issueTokenFeatures,
-      [event.target.name]: event.target.checked,
-    });
-    //console.log(`features ${JSON.stringify(issueTokenFeatures)}`)
-  };
 
   const handleInputChange = (e: any) => {
     const { name, value } = e.target;
     setFormData((prev) => ({ ...prev, [name]: value }));
   };
 
-  const handleTxChange = (e: any) => {
-    const { name, value } = e.target;
-    setTxTransferForm((prev) => ({ ...prev, [name]: value }));
-  };
-
-  const handleTxIssueChange = (e: any) => {
-    const { name, value } = e.target;
-    setTxIssueForm((prev) => ({ ...prev, [name]: value }));
-  };
-  const handleAddMsg = () => {
-    // validate balance
-    // validate destination address
-    // create formatted msg for proposal
-    let msg = {
-      bank: {
-        send : {
-          from_address: vaultSelected,
-          to_address: txTransferForm.destination,
-          amount: [{denom: vaultFtSelected?.denom, amount: txTransferForm.amount}]
-        }
-      }
-    }
-    setMsgs((msgs) => [...msgs, msg])
-  }
-
-  // WIP
-  const handleAddIssueMsg = () => {
-    let features = []
-    if (issueTokenFeatures["mintable"]) {
-      features.push(1)
-    } 
-    if (issueTokenFeatures["burnable"]) {
-      features.push(2)
-    }
-    // try with wasm: { instantiate: {MsgInstantiate}}
-    const msgIssueFT = {
-          typeUrl: "/coreum.asset.ft.v1.MsgIssue",  // error (expected "custom", "bank" or "wasm")
-          value: MsgIssue.fromPartial({
-            issuer: vaultSelected,
-            subunit: txIssueForm.subunit,
-            symbol: txIssueForm.symbol,
-            precision: txIssueForm.precision,
-            initialAmount: txIssueForm.initialAmount,
-            features: features,//txIssueForm.features,
-            sendCommissionRate: txIssueForm.sendCommissionRate// `${Decimal.fromUserInput("0.5", 18).atomics}` / // 50% input
-          }),
-    };
-    let te = new TextEncoder()
-    let msg_e = te.encode(JSON.stringify(msgIssueFT))
-    let msg = {
-      custom: {
-        msg: msg_e
-      }
-    }
-
-    setMsgs((msgs) => [...msgs, msg])
-  }
 
   async function createProposal() {
     if (!vaults || !signingClient) return
@@ -154,7 +60,10 @@ const Proposal: NextPage = () => {
       setSuccess("your proposal has been created!")
     }
   }
-
+  function addMsg(m: any) {
+    console.log(`enters ${JSON.stringify(m)}`)
+    setMsgs((msgs) => [...msgs, m])
+  }
   async function getBalancesFT(vaultAddress: string) {
     for (let t = 0; t < tokensList.length; t++) {
       let tokenDenom = tokensList[t]
@@ -168,7 +77,6 @@ const Proposal: NextPage = () => {
   }
   useEffect(() => {
     if (vaultSelected) {
-      setVaultFtSelected(null)
       setVaultBalance([])
       getBalancesFT(vaultSelected)
     }
@@ -194,170 +102,20 @@ const Proposal: NextPage = () => {
       value={formData.description}
       onChange={handleInputChange}
     />
-    <hr />
-    Add txs <br />
+
+    {' '}from vault:{' '} 
     <Select
-      value={txModel}
-      onChange={(e) => setTxModel(e.target.value)}
+      value={vaultSelected}
+      onChange={(e) => setVaultSelected(e.target.value)}
     >
-      <MenuItem value={'none'}>None</MenuItem>
-      <MenuItem value={'send-ft'}>Send FT</MenuItem>
-      <MenuItem value={'send-nft'}>Send NFT</MenuItem>
-      <MenuItem value={'create-ft'}>Create FT</MenuItem>
-      <MenuItem value={'create-nft'}>Create NFT</MenuItem>
-      <MenuItem disabled value={'members'}>Add/Remove Member</MenuItem>
+      {vaults?.map((v) => {return(<MenuItem value={v}>{v}</MenuItem>)})}
     </Select>
-    {txModel === 'create-ft' &&
-      <>
-      {' '}from vault:{' '} 
-      <Select
-        value={vaultSelected}
-        onChange={(e) => setVaultSelected(e.target.value)}
-      >
-        {vaults?.map((v) => {return(<MenuItem value={v}>{v}</MenuItem>)})}
-      </Select>
-      <br />
-
-      <TextField
-        margin="normal"
-        label="Subunit"
-        name="subunit"
-        value={txIssueForm.subunit}
-        onChange={handleTxIssueChange}
-      />
-      <TextField
-        margin="normal"
-        label="Token Symbol"
-        name="symbol"
-        value={txIssueForm.symbol}
-        onChange={handleTxIssueChange}
-      />
-      <TextField
-        margin="normal"
-        label="Initial amount"
-        name="initialAmount"
-        placeholder="Supply (e.g., 1,000,000)"
-        value={txIssueForm.initialAmount}
-        onChange={handleTxIssueChange}
-      />
-      <TextField
-        margin="normal"
-        label="Precision"
-        name="decimals"
-        type="number"
-        sx={{ width: '8rem' }}
-        value={txIssueForm.precision}
-        onChange={handleTxIssueChange}
-        inputProps={{ min: 0, max: 18 }} // This constrains input to the range 0-18
-      />
-      <TextField
-        margin="normal"
-        label="Send comission"
-        name="sendCommissionRate"
-        type="number"
-        sx={{ width: '8rem' }}
-        value={txIssueForm.sendCommissionRate}
-        onChange={handleTxIssueChange}
-        inputProps={{ min: 0, max: 100 }}
-        />
-        
-      <FormControl component="fieldset">
-        <FormLabel component="legend">Token features</FormLabel>
-        <FormGroup aria-label="position" row>
-          <FormControlLabel
-            value="Mintable"
-            control={<Checkbox
-                checked={issueTokenFeatures.mintable}
-                onChange={handleFeaturesChange}
-                name="mintable"
-              />}
-            label="Mintable"
-            labelPlacement="top"
-          />
-          <FormControlLabel
-            value="Burnable"
-            control={<Checkbox
-                checked={issueTokenFeatures.burnable}
-                onChange={handleFeaturesChange}
-                name="burnable"
-              />}
-            label="Burnable"
-            labelPlacement="top"
-          />
-          <FormControlLabel
-            value="Freezable"
-            disabled
-            control={<Checkbox />}
-            label="Freezable"
-            labelPlacement="top"
-          />
-          <FormControlLabel
-            value="Unfreezable"
-            disabled
-            control={<Checkbox />}
-            label="Unfreezable"
-            labelPlacement="top"
-          />
-        </FormGroup>
-      </FormControl>
-
-      <Button
-        onClick={handleAddIssueMsg}
-      >Add tx</Button>
-      </>
-    }
-    {txModel === 'create-nft' &&
-      <>Create NFT from your guild</>
-    }
-
-    {txModel === 'send-ft' &&
-      <>
-        <Select
-          value={vaultSelected}
-          onChange={(e) => setVaultSelected(e.target.value)}
-        >
-          {vaults?.map((v) => {return(<MenuItem value={v}>{v}</MenuItem>)})}
-        </Select>
-        {vaultSelected && vaultBalance &&
-          <>
-          <Select
-            value={vaultFtSelected}
-            onChange={(t) => setVaultFtSelected(t.target.value)}
-          >
-            {vaultBalance.map((b) => {return( <MenuItem value={b}>{b.denom}</MenuItem>)})}
-          </Select>
-          {vaultFtSelected &&
-            <>
-              Max: {vaultFtSelected.amount}
-              <hr />
-              <TextField        
-                margin="normal"
-                label="Destination"
-                name="destination"
-                value={txTransferForm.destination}
-                onChange={handleTxChange}
-              />
-              <TextField
-                margin="normal"
-                label="amount"
-                name="amount"
-                type='number'
-                inputProps={{min: 0, max: vaultFtSelected.amount}}                
-                value={txTransferForm.amount}
-                onChange={handleTxChange}
-              >Amount</TextField>
-              <Button
-                onClick={handleAddMsg}
-              >Add tx</Button>
-            </>
-          }
-          </>
-        }
-      </>
-    }
-    {txModel === 'send-nft' &&
-      <>Select vault, choose NFT, select destination</>
-    }
+    <hr />
+    <ProposalTxCreator 
+      vaultSelected={vaultSelected}
+      vaultBalance={vaultBalance}
+      setMsgs={(m)=>addMsg(m)} 
+    />
     <hr />
     <Button
       style={{marginTop: "50px"}}
